requirements.txt
Django>=5.0.0
django-ninja
psycopg2-binary
python-dotenv
vnstock
PyJWT>=2.8.0
google-auth>=2.20.0
requests>=2.31.0
python==3.11.1

